refactor database init things for pytest into helpers so fastapi pkgs can use them too
